+++ packages/veui/src/components/Calendar/index.js
@@ -1,0 +1,2 @@
+import Calendar from './Calendar'
+export default Calendar